rename csi to storage
