--- csi/doc.go
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright (c) 2022 Dell Inc, or its subsidiaries.
-
-// Package csi implements the go library for OPI to be used in CSI drivers
-package csi
