+++ storage/doc.go
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (c) 2022 Dell Inc, or its subsidiaries.
+
+// Package storage implements the go library for OPI to be used in storage, for example, CSI drivers
+package storage
