+++ storage/goopicsi.go
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) 2022 Dell Inc, or its subsidiaries.
 
-// Package csi implements the go library for OPI to be used in CSI drivers
-package csi
+// Package storage implements the go library for OPI to be used in storage, for example, CSI drivers
+package storage
 
 import (
 	"context"
